requirements.txt
streamlit
qrcode
pillow
wordcloud
matplotlib